The ability to reduce redirect hops on domain redirects
- Domain redirects were in place and if an original request came in on
http it would first redirect to the correct domain using the same
protocol '$schema' as defined in the nginx config, but because the site
was https everywhere it would then get an additional redirect over to
the secure https url. This template change allows for an optional
argument to be set on the domain redirect to override the protocol
used on the redirect while maintaining the default of '$schema' if a
specific protocol is not set.